FMS, DialogFenster, Szenenstruktur, Dialog
+ Die Player-Instanz kann über die Hauptklasse "Game.Instance.Player"
erreicht werden!
+ Player: "Start" gegen "Awake" augetauscht
+ Player Bewegungen können bei Dialogen abgeschaltet werden
+ Antworten werden bei Zustandswechsel immer gelöscht!
+ Antwortattribute "nextText" wird in Dialogen unterstützt (Ringaufrufe
zurzeit nicht unterstützt!)
+ Zustandsgraph bis zu der Erklärung des Springens fertig
+ Neuer FSM Zustand: S_Empty
- GameObject gelöscht "DialogWindow" (wird dynamisch generiert)
- Den einfachen Dialog entfernt
# Bugfixes in "DialogueWindowScript"

+++ Assets/Scripts/Game/States/Game/Dialogue/Advanced/AdvancedDialogueParser.cs
@@ -203,7 +203,7 @@ public abstract class AdvancedDialogueParser
         // Local variables
         string text = "";
         string character = "";
-        int timeInSec = -1;
+        double timeInSec = -1;
         TextPart part = null;
         XmlNode xmlNodeAtt = null;
 
@@ -214,7 +214,7 @@ public abstract class AdvancedDialogueParser
             Debug.LogError("Text-parts need a defined display time (in seconds)!");
             return null;
         }
-        timeInSec = int.Parse(xmlNodeAtt.Value);
+        timeInSec = double.Parse(xmlNodeAtt.Value);
 
         // Get attribute: character
         xmlNodeAtt = _node.Attributes.GetNamedItem("character");
